move progress bar to it's own file
